requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
streamlit==1.51.0
nest_asyncio==1.6.0
coloredlogs==15.0.1
pydantic==2.12.4
langchain_core==1.1.0
langchain_community==0.4.1
ipython==9.7.0
requests==2.32.5
langchain_openai==1.0.3
langchain_experimental==0.4.0
pillow==12.0.0
langchain-huggingface==1.0.1
huggingface_hub==0.36.0
python-dotenv==1.2.1
langchain-astradb==0.6.0
open_clip_torch==3.2.0
torch==2.9.1
python-multipart==0.0.20
langchain-qdrant==1.1.0
structlog==25.4.0
